backend/requirements.txt
# Main dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
email-validator==2.1.0.post1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
neo4j==5.14.1

aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
python-ldap==3.4.0
hl7apy==1.3.5
aiofiles==23.2.1
prometheus-fastapi-instrumentator==6.0.0
sentry-sdk==1.39.2
python-socketio==5.12.0
itsdangerous==2.1.2
starlette_prometheus==0.10.0
watchdog==3.0.0

# Test dependencies
pytest==8.4.1
pytest-cov==6.2.1
pytest-asyncio==1.0.0
pytest-mock==3.14.1
pytest-xdist==3.6.1
pytest-env==1.1.3
pytest-sugar==1.0.0
pytest-factoryboy==2.6.1
pytest-freezegun==0.4.2
